change app logic

+++ js/src/App.js
@@ -6,15 +6,23 @@ import UntrustedClient from "./UntrustedClient";
 import KeyPair from "./KeyPair";
 
 import TorrentClient from "./TorrentClient";
+import TorrentDownloader from "./TorrentDownloader";
 
 async function fetchPeerData() {
   const SAMPLESITE =
     "magnet:?xt=urn:btih:e60f82343019bd711c5c731b46e118b0f2b2ecc6&dn=samplesite&tr=ws%3A%2F%2Flocalhost%3A4444&tr=wss%3A%2F%2Ftracker.btorrent.xyz&tr=wss%3A%2F%2Ftracker.fastcast.nz&tr=wss%3A%2F%2Ftracker.openwebtorrent.com";
-  let client = new TorrentClient();
-  client.verbose = true;
-  let torrent = await client.download(SAMPLESITE);
-  await torrent.monitorProgress();
-  client.destroy();
+
+  if (false) {
+    let client = new TorrentClient();
+    client.verbose = true;
+    let torrent = await client.download(SAMPLESITE);
+    await torrent.monitorProgress();
+    client.destroy();
+  } else {
+    let downloader = new TorrentDownloader();
+    await downloader.downloadMagnet(SAMPLESITE);
+    console.log("downloadMagnet done");
+  }
 }
 
 export default class App extends Component {
